Refactor utilities into a centralized utils module
Moved and refactored various utility functions to a centralized utils module for better organization and reusability. Updated imports in multiclass_utils for backward compatibility, and ensured seamless integration with prior tests and modules.
