+++ pytorch_tabnet/multiclass_utils/__init__.py
@@ -1,3 +1,3 @@
-from .dimension import infer_multitask_output as infer_multitask_output
-from .dimension import infer_output_dim as infer_output_dim
+from ..utils.dimension import infer_multitask_output as infer_multitask_output
+from ..utils.dimension import infer_output_dim as infer_output_dim
 from .validation import check_output_dim as check_output_dim
