+++ pytorch_tabnet/multiclass_utils/_is_integral_float.py
@@ -1,5 +1,6 @@
-import numpy as np
+"""Import from utils module for backward compatibility."""
 
+from pytorch_tabnet.utils._is_integral_float import _is_integral_float  # noqa
 
-def _is_integral_float(y: np.ndarray) -> bool:
-    return y.dtype.kind == "f" and np.all(y.astype(int) == y)
+# from ..utils._is_integral_float import is_integral_float
+# _is_integral_float
