+++ pytorch_tabnet/utils/__init__.py
@@ -1,8 +1,11 @@
 """Utility functions for TabNet package."""
 
 from .device import define_device as define_device
+from .dimension import infer_multitask_output as infer_multitask_output
+from .dimension import infer_output_dim as infer_output_dim
 from .matrices import create_explain_matrix as create_explain_matrix
 from .matrices import create_group_matrix as create_group_matrix
+from .multiclass_validation import check_output_dim as check_output_dim
 from .serialization import ComplexEncoder as ComplexEncoder
 from .validation import check_embedding_parameters as check_embedding_parameters
 from .validation import check_input as check_input
